Fixed various warnings, added support for 512x512 pixel icons, fixed support for 256x256 pixel icons, changed SDK version to 10.6.

+++ CSIconViewItem.m
@@ -158,7 +158,7 @@
 - (void)setLabelColor:(NSColor *)newColor
 {
   if (labelColor != newColor) {
-    float L, a, b, alpha;
+    CGFloat L, a, b, alpha;
     
     NSColor *oldColor = labelColor;
     labelColor = [newColor retain];
